gcp/pyproject.toml
[project]
name = "gcp"
version = "0.1.0"
description = ""
authors = [
    {name = "uuboyscy",email = "aegis12321@gmail.com"}
]
readme = "README.md"
requires-python = "3.12.4"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "google-cloud-bigquery (>=3.29.0,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "db-dtypes (>=1.4.1,<2.0.0)",
    "flask (>=3.1.0,<4.0.0)",
    "google-cloud-secret-manager (>=2.23.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
